refactor/#175: issueService 및 IssueAcceptanceTest 분리
