+++ BE/src/test/java/com/issuetracker/util/steps/IssueSteps.java
@@ -10,8 +10,6 @@ import com.issuetracker.issue.ui.dto.IssueCreateRequest;
 import com.issuetracker.issue.ui.dto.IssueSearchRequest;
 import com.issuetracker.issue.ui.dto.IssueUpdateAllOpenRequest;
 import com.issuetracker.issue.ui.dto.IssueUpdateRequest;
-import com.issuetracker.issue.ui.dto.assignedlabel.AssignedLabelCreateRequest;
-import com.issuetracker.issue.ui.dto.assignee.AssigneeCreateRequest;
 import com.issuetracker.issue.ui.dto.comment.IssueCommentCreateRequest;
 
 import io.restassured.RestAssured;
@@ -21,7 +19,7 @@ import io.restassured.response.Response;
 public class IssueSteps {
 
 	public static ExtractableResponse<Response> 이슈_목록_조회_요청(IssueSearchRequest issueSearchRequest) {
-		Map<String, Object> params =  new HashMap<>();
+		Map<String, Object> params = new HashMap<>();
 		params.put("isOpen", issueSearchRequest.getIsOpen());
 		params.put("assigneeIds", issueSearchRequest.getAssigneeIds());
 		params.put("labelIds", issueSearchRequest.getLabelIds());
@@ -57,26 +55,12 @@ public class IssueSteps {
 			.accept(MediaType.APPLICATION_JSON_VALUE)
 			.when().get("/api/issues/milestones")
 			.then().log().all().extract();
-  	}
-  
+	}
+
 	public static ExtractableResponse<Response> 작성자_목록_조회_요청() {
 		return RestAssured.given().log().all()
 			.accept(MediaType.APPLICATION_JSON_VALUE)
 			.when().get("/api/issues/authors")
-			.then().log().all().extract();
-  	}
-  
-	public static ExtractableResponse<Response> 이슈에_등록_되어있는_담당자_목록_조회_요청() {
-		return RestAssured.given().log().all()
-			.accept(MediaType.APPLICATION_JSON_VALUE)
-			.when().get("/api/issues/assignees")
-			.then().log().all().extract();
-	}
-
-	public static ExtractableResponse<Response> 이슈에_등록_되어있는_라벨_목록_조회_요청() {
-		return RestAssured.given().log().all()
-			.accept(MediaType.APPLICATION_JSON_VALUE)
-			.when().get("/api/issues/labels")
 			.then().log().all().extract();
 	}
 
@@ -154,52 +138,6 @@ public class IssueSteps {
 			.contentType(MediaType.APPLICATION_JSON_VALUE)
 			.accept(MediaType.APPLICATION_JSON_VALUE)
 			.when().patch("/api/issues/{id}/comments/{comment-id}", id, issueCommentId)
-			.then().log().all().extract();
-	}
-
-	public static ExtractableResponse<Response> 이슈에_등록_및_삭제될_담당자_목록_조회_요청(Long id) {
-		return RestAssured.given().log().all()
-			.accept(MediaType.APPLICATION_JSON_VALUE)
-			.when().get("/api/issues/{id}/assignee-candidates", id)
-			.then().log().all().extract();
-	}
-
-	public static ExtractableResponse<Response> 이슈에_담당자_등록_요청(Long id, Long memberId) {
-		return RestAssured.given().log().all()
-			.body(new AssigneeCreateRequest(memberId))
-			.contentType(MediaType.APPLICATION_JSON_VALUE)
-			.accept(MediaType.APPLICATION_JSON_VALUE)
-			.when().post("/api/issues/{id}/assignees", id)
-			.then().log().all().extract();
-	}
-
-	public static ExtractableResponse<Response> 이슈에_담당자_삭제_요청(Long id, Long assigneeId) {
-		return RestAssured.given().log().all()
-			.accept(MediaType.APPLICATION_JSON_VALUE)
-			.when().delete("/api/issues/{id}/assignees/{assignee-id}", id, assigneeId)
-			.then().log().all().extract();
-	}
-
-	public static ExtractableResponse<Response> 이슈에_등록_및_삭제될_라벨_목록_조회_요청(Long id) {
-		return RestAssured.given().log().all()
-			.accept(MediaType.APPLICATION_JSON_VALUE)
-			.when().get("/api/issues/{id}/label-candidates", id)
-			.then().log().all().extract();
-	}
-
-	public static ExtractableResponse<Response> 이슈에_라벨_등록_요청(Long id, Long labelId) {
-		return RestAssured.given().log().all()
-			.body(new AssignedLabelCreateRequest(labelId))
-			.contentType(MediaType.APPLICATION_JSON_VALUE)
-			.accept(MediaType.APPLICATION_JSON_VALUE)
-			.when().post("/api/issues/{id}/assigned-labels", id)
-			.then().log().all().extract();
-	}
-
-	public static ExtractableResponse<Response> 이슈에_라벨_삭제_요청(Long id, Long assignedLabelId) {
-		return RestAssured.given().log().all()
-			.accept(MediaType.APPLICATION_JSON_VALUE)
-			.when().delete("/api/issues/{id}/assigned-labels/{assigned-label-id}", id, assignedLabelId)
 			.then().log().all().extract();
 	}
 }
